Add mobile module and update dependencies
A new module 'mobile' has been added to the project. It has also been added as a dependency in the 'app' and 'sampler' modules. The stylesheets for 'mobile' have been linked in the 'app' and 'sampler' modules. Add 'mobile' to the list of project modules in the 'pom.xml' file.