Add LotwPageView and HelloLotwPageView classes
This commit adds two new classes, LotwPageView.java and HelloLotwPageView.java. LotwPageView provides functionalities to display links of the week, enabling sorting, and displaying based on creation dates. It also includes the addition of necessary style properties in the mobile.css file. HelloLotwPageView is a demo class displaying an instance of LotwPageView.