Update click handling in list and hyperlinks
Refactored the event handlers for clicking on list items and hyperlinks. Clicks are now only activated when the mouse has not moved between the press and release events. A new ListViewUtil class has been created to encapsulate this logic for list items, cleaning up code in other areas of the application.