Add search functionality to mobile version
A new Search page was added to the mobile version of the application. To accommodate this, the search bar on the home page was modified to redirect to the Search page when clicked. Additional enhancements include updated styling for better UI experience and automatic focusing on the search bar when the Search page is opened.