Refactor MobileCategoryPageBase and MobileDetailsPageBase to extend MobilePageBase
This update involves refactoring various mobile components such as MobileCategoryPageBase and MobileDetailsPageBase to inherit from a new MobilePageBase class which replaces the older SizeSupport logic. Updated MobileRouter to handle new type of MobileResponse, replaced the viewCallback() with new responseCallback() in MobileRoute. Added a new utility class MobileResponse to manage the response data from the mobile route and a new class MobilePageBase to handle all the sizing loic and also manage the swipe interactions and callbacks.